:recycle: FIX: Blob 함수분리
